chore(utxo-staking): move files into coredao specific folder
TICKET: BTC-1579
