+++ modules/utxo-staking/src/coreDao/descriptor.ts
