+++ modules/utxo-staking/src/coreDao/index.ts
@@ -1,0 +1,2 @@
+export * from './opReturn';
+export * from './descriptor';
