+++ modules/utxo-staking/test/fixtures/coreDao/descriptor/sh-2of2-asm.txt
