+++ modules/utxo-staking/test/fixtures/coreDao/descriptor/sh-3of3-tx.txt
