+++ modules/utxo-staking/test/fixtures/coreDao/descriptor/sh-wsh-2of2-string.txt
